Resolve merging issues with the BE
