+++ backend/src/main/java/com/example/catSpringBoot/FullstackCatSpringbootApplication.java
@@ -1,0 +1,31 @@
+package com.example.catSpringBoot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@SpringBootApplication
+@EnableJpaRepositories
+@EnableConfigurationProperties
+@EntityScan
+public class FullstackCatSpringbootApplication{
+    public static void main(String[] args){
+        SpringApplication.run(FullstackCatSpringbootApplication.class,args);
+    }
+
+    @Bean
+    public WebMvcConfigurer corsConfigurer() {
+        return new WebMvcConfigurer() {
+            @Override
+            public void addCorsMappings(CorsRegistry registry) {
+                registry.addMapping("/**").allowedMethods("GET", "PUT", "POST", "DELETE").allowedOrigins("http://localhost:3000");
+            }
+        };
+    }
+    
+}
