+++ backend/src/main/java/com/example/catSpringBoot/config/AppConfig.java
@@ -1,0 +1,16 @@
+package com.example.catSpringBoot.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import com.example.catSpringBoot.repository.CatRepository;
+import com.example.catSpringBoot.service.CatService;
+
+@Configuration
+public class AppConfig {
+    private CatRepository catRepository;
+    @Bean
+    public CatService CatService(){
+        return new CatService(this.catRepository);
+    }
+}
