+++ backend/src/main/java/com/example/catSpringBoot/config/WebMvcConfig.java
@@ -1,0 +1,18 @@
+package com.example.catSpringBoot.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedMethods("GET", "PUT", "POST", "DELETE")
+                .allowedOrigins("http://localhost:3000")
+                .allowCredentials(true)
+                .allowedHeaders("*");
+    }
+}
