+++ backend/src/main/java/com/example/catSpringBoot/dto/CreateCatDto.java
@@ -1,0 +1,11 @@
+package com.example.catSpringBoot.dto;
+
+import lombok.Data;
+
+@Data
+public class CreateCatDto {
+    private String name;
+    private String breed;
+    private String description;
+    private String image;
+}
