+++ backend/src/main/java/com/example/catSpringBoot/dto/ErrorDto.java
@@ -1,0 +1,10 @@
+package com.example.catSpringBoot.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ErrorDto {
+    private String msg;
+}
