+++ backend/src/main/java/com/example/catSpringBoot/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.example.catSpringBoot.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.catSpringBoot.model.User;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUsername(String username);
+}
