--- backend/src/main/java/com/example/fullstackbooktodospringboot/FullstackbookTodoSpringbootApplication.java
@@ -1,15 +1,0 @@
-package com.example.fullstackbooktodospringboot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.web.servlet.ServletComponentScan;
-
-@ServletComponentScan
-@SpringBootApplication
-public class FullstackbookTodoSpringbootApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(FullstackbookTodoSpringbootApplication.class, args);
-    }
-
-}
