--- backend/src/main/java/com/example/fullstackbooktodospringboot/config/WebMvcConfig.java
@@ -1,18 +1,0 @@
-package com.example.fullstackbooktodospringboot.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedMethods("GET", "PUT", "POST", "DELETE")
-                .allowedOrigins("http://localhost:3000")
-                .allowCredentials(true)
-                .allowedHeaders("*");
-    }
-}
