--- backend/src/main/java/com/example/fullstackbooktodospringboot/controller/ToDoController.java
@@ -1,55 +1,0 @@
-package com.example.fullstackbooktodospringboot.controller;
-
-import com.example.fullstackbooktodospringboot.dto.CreateToDoDto;
-import com.example.fullstackbooktodospringboot.dto.ToDoDto;
-import com.example.fullstackbooktodospringboot.dto.UpdateToDoDto;
-import com.example.fullstackbooktodospringboot.service.ToDoService;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.Optional;
-
-@RestController
-@RequestMapping("/todos")
-public class ToDoController {
-    private ToDoService toDoService;
-
-    public ToDoController(ToDoService toDoService) {
-        this.toDoService = toDoService;
-    }
-
-    @PostMapping("")
-    public ResponseEntity<ToDoDto> createToDo(@RequestBody CreateToDoDto newToDo) {
-        ToDoDto toDoDTO = toDoService.createTodo(newToDo);
-        return new ResponseEntity<>(toDoDTO, HttpStatus.CREATED);
-    }
-
-    @GetMapping("")
-    public List<ToDoDto> getToDos(@RequestParam Optional<Boolean> completed, Authentication authentication) {
-        // String username = authentication.getName();
-        // System.out.println("Username2: " + username);
-        if (completed.isPresent()) {
-            return toDoService.getToDos(completed.get());
-        }
-        return toDoService.getToDos();
-    }
-
-    @GetMapping("/{id}")
-    public ToDoDto getToDoById(@PathVariable Long id) {
-        return toDoService.getToDoById(id);
-    }
-
-    @PutMapping("/{id}")
-    public ToDoDto updateToDo(@PathVariable Long id, @RequestBody UpdateToDoDto updateToDoDto) {
-        return toDoService.updateToDo(id, updateToDoDto);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> deleteToDo(@PathVariable Long id) {
-        toDoService.deleteToDo(id);
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
-}
