--- backend/src/main/java/com/example/fullstackbooktodospringboot/dto/CreateToDoDto.java
@@ -1,9 +1,0 @@
-package com.example.fullstackbooktodospringboot.dto;
-
-import lombok.Data;
-
-@Data
-public class CreateToDoDto {
-    private String name;
-    private Boolean completed;
-}
