--- backend/src/main/java/com/example/fullstackbooktodospringboot/dto/ErrorDto.java
@@ -1,10 +1,0 @@
-package com.example.fullstackbooktodospringboot.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ErrorDto {
-    private String msg;
-}
