--- backend/src/main/java/com/example/fullstackbooktodospringboot/dto/ToDoDto.java
@@ -1,19 +1,0 @@
-package com.example.fullstackbooktodospringboot.dto;
-
-import com.example.fullstackbooktodospringboot.model.ToDo;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ToDoDto {
-    private Long id;
-    private String name;
-    private Boolean completed;
-
-    public ToDoDto(ToDo entity) {
-        this.id = entity.getId();
-        this.name = entity.getName();
-        this.completed = entity.getCompleted();
-    }
-}
