--- backend/src/main/java/com/example/fullstackbooktodospringboot/model/Cat.java
@@ -1,20 +1,0 @@
-package com.example.fullstackbooktodospringboot.model;
-
-import lombok.Data;
-
-import javax.persistence.*;
-
-@Entity
-@Data
-@Table(name = "todos")
-public class Cat {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column
-    private String name;
-
-    @Column
-    private Boolean completed;
-}
