--- backend/src/main/java/com/example/fullstackbooktodospringboot/model/ToDo.java
@@ -1,20 +1,0 @@
-package com.example.fullstackbooktodospringboot.model;
-
-import lombok.Data;
-
-import javax.persistence.*;
-
-@Entity
-@Data
-@Table(name = "todos")
-public class ToDo {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column
-    private String name;
-
-    @Column
-    private Boolean completed;
-}
