--- backend/src/main/java/com/example/fullstackbooktodospringboot/repository/ToDoRepository.java
@@ -1,10 +1,0 @@
-package com.example.fullstackbooktodospringboot.repository;
-
-import com.example.fullstackbooktodospringboot.model.ToDo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface ToDoRepository extends JpaRepository<ToDo, Long> {
-    List<ToDo> findByCompleted(Boolean completed);
-}
