--- backend/src/main/java/com/example/fullstackbooktodospringboot/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.example.fullstackbooktodospringboot.repository;
-
-import com.example.fullstackbooktodospringboot.model.User;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-}
