--- backend/src/main/java/com/example/fullstackbooktodospringboot/security/TokenAuthenticationFilter.java
@@ -1,53 +1,0 @@
-package com.example.fullstackbooktodospringboot.security;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class TokenAuthenticationFilter extends OncePerRequestFilter {
-
-    @Autowired
-    private TokenProvider tokenProvider;
-
-    private static final Logger logger = LoggerFactory.getLogger(TokenAuthenticationFilter.class);
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-            throws ServletException, IOException {
-        try {
-            // need to implement handling for when token is not valid or expired, when a
-            // invalid token is sent or no token is sent the exception is thrown
-            String token = resolveToken(request);
-            if (StringUtils.hasText(token) && tokenProvider.validateToken(token)) {
-                String username = tokenProvider.getUsernameFromToken(token);
-                // System.out.println("Username1: " + username);
-                Authentication authentication = new UsernamePasswordAuthenticationToken(username, null, null);
-                SecurityContextHolder.getContext().setAuthentication(authentication);
-            }
-        } catch (Exception e) {
-            logger.error("Could not set user authentication in security context", e);
-        }
-        filterChain.doFilter(request, response);
-    }
-
-    private String resolveToken(HttpServletRequest request) {
-        String bearerToken = request.getHeader("Authorization");
-        // System.out.println("Bearer token: " + bearerToken);
-        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring(7, bearerToken.length());
-        }
-        return null;
-    }
-}
