+++ backend/src/main/resources/application.properties
@@ -3,8 +3,9 @@ spring.datasource.url=${DB_URL:jdbc:postgresql://localhost:80/postgres}
 spring.datasource.username=${DB_USER:postgres}
 spring.datasource.password=${DB_PASSWORD:postgres}
 server.port=5432
-
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.generate-ddl=true
+spring.jpa.properties.hibernate.hbm2dll.create_namespaces=true
+spring.jpa.hibernate.ddl-auto = create
 
 spring.security.oauth2.client.registration.github.client-id=af39404ec65880c471ce
 spring.security.oauth2.client.registration.github.client-secret=e88f2cdbf74464cd4566d4e125b3e541cd4eb207
