--- backend/src/test/java/com/example/fullstackbooktodospringboot/FullstackbookTodoSpringbootApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.fullstackbooktodospringboot;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class FullstackbookTodoSpringbootApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
